+++ packages/Bio/src/tests/renderers-monomer-placer-tests.ts
@@ -8,8 +8,7 @@ import {category, test} from '@datagrok-libraries/utils/src/test';
 import {MonomerPlacer} from '@datagrok-libraries/bio/src/utils/cell-renderer-monomer-placer';
 import {monomerToShort} from '@datagrok-libraries/bio/src/utils/macromolecule';
 import {SeqHandler} from '@datagrok-libraries/bio/src/utils/seq-handler';
-
-import {MonomerLibManager} from '../utils/monomer-lib/lib-manager';
+import {getMonomerLibHelper} from '@datagrok-libraries/bio/src/monomer-works/monomer-utils';
 
 import {_package} from '../package-test';
 
@@ -45,7 +44,7 @@ category('renderers: monomerPlacer', () => {
         'id1,m1-M-m3-mon4-mon5-N-T-MON8-N9\n' +
         'id2,m1-mon2-m3-mon4-mon5-Num--MON8-N9\n' +
         'id3,\n' + // empty
-        'id4,mon1-M-mon3-mon4-mon5---MON8-N9\n',
+        'id4,mon1-M-mon3-mon4-mon5---MON8-N9\n', // [ 5, 38, 71, 104, 137, 170, 203, 236, 269, 295 ]
       testList: [
         {src: {row: 0, x: -1}, tgt: {pos: null}},
         {src: {row: 1, x: 0}, tgt: {pos: null}},
@@ -59,7 +58,7 @@ category('renderers: monomerPlacer', () => {
         {src: {row: 2, x: 20}, tgt: {pos: null}}, // empty value
         {src: {row: 3, x: 170}, tgt: {pos: 4}},
         {src: {row: 3, x: 200}, tgt: {pos: 5}},
-        {src: {row: 3, x: 282}, tgt: {pos: null}},
+        {src: {row: 3, x: 297}, tgt: {pos: null}},
       ]
     },
     fastaMsa: {
@@ -88,6 +87,8 @@ id3,QHIRE--LT
 
   for (const [testName, testData] of Object.entries(tests)) {
     test(`getPosition-${testName}`, async () => {
+      const libHelper = await getMonomerLibHelper();
+      const monomerLib = libHelper.getBioLib();
       const df: DG.DataFrame = DG.DataFrame.fromCsv(testData.csv);
       await grok.data.detectSemanticTypes(df);
       const seqCol: DG.Column = df.getCol('seq');
@@ -103,18 +104,18 @@ id3,QHIRE--LT
           separatorWidth: sepWidth,
           monomerToShort: monomerToShort,
           monomerLengthLimit: monLength,
-          monomerLib: MonomerLibManager.instance.getBioLib(),
         };
       });
 
+      const width: number = 10000;
       const testList = testData.testList;
       // simulate rendering
       for (let rowI: number = 0; rowI < seqCol.length; ++rowI)
-        colTemp.getCellMonomerLengths(rowI);
+        colTemp.getCellMonomerLengths(rowI, 10000);
 
       const errorList: string[] = [];
       for (const [test, _testI] of wu.enumerate(testList)) {
-        const res = {pos: colTemp.getPosition(test.src.row, test.src.x)};
+        const res = {pos: colTemp.getPosition(test.src.row, test.src.x, width)};
         if (test.tgt.pos != res.pos) {
           errorList.push(`Test src ${JSON.stringify(test.src)} expected tgt ${JSON.stringify(test.tgt)},` +
             ` but get ${JSON.stringify({res})}`);