+++ packages/Bio/src/utils/helm-to-molfile/converter/converter.ts
@@ -5,9 +5,12 @@ import * as OCL from 'openchemlib/full';
 
 import {errInfo} from '@datagrok-libraries/bio/src/utils/err-info';
 import {RDModule, RDMol} from '@datagrok-libraries/chem-meta/src/rdkit-api';
-import {_package} from '../../../package';
+import {IMonomerLib} from '@datagrok-libraries/bio/src/types/index';
+
 import {Polymer} from './polymer';
 import {GlobalMonomerPositionHandler} from './position-handler';
+
+import {_package, getMonomerLibHelper} from '../../../package';
 
 export class HelmToMolfileConverter {
   constructor(private helmColumn: DG.Column<string>, private df: DG.DataFrame) {
@@ -54,7 +57,7 @@ export class HelmToMolfileConverter {
   async convertToRdKitBeautifiedMolfileColumn(chiralityEngine?: boolean): Promise<DG.Column<string>> {
     const molfilesV3K = (await this.convertToMolfileV3KColumn()).toList();
     const rdKitModule: RDModule = await grok.functions.call('Chem:getRdKitModule');
-    const beautifiedMols = molfilesV3K.map((item) =>{
+    const beautifiedMols = molfilesV3K.map((item) => {
       if (item === '')
         return null;
       const mol = rdKitModule.get_mol(item);
@@ -81,6 +84,7 @@ export class HelmToMolfileConverter {
   private async convertToMolfileV3KColumn(): Promise<DG.Column<string>> {
     const polymerGraphColumn: DG.Column<string> = await this.getPolymerGraphColumn();
     const rdKitModule = await grok.functions.call('Chem:getRdKitModule');
+    const monomerLib: IMonomerLib = (await getMonomerLibHelper()).getBioLib();
     const molfileList = polymerGraphColumn.toList().map(
       (pseudoMolfile: string, idx: number) => {
         const helm = this.helmColumn.get(idx);
@@ -88,7 +92,7 @@ export class HelmToMolfileConverter {
           return '';
         let result = '';
         try {
-          result = this.getPolymerMolfile(helm, pseudoMolfile, rdKitModule);
+          result = this.getPolymerMolfile(helm, pseudoMolfile, rdKitModule, monomerLib);
         } catch (err: any) {
           const [errMsg, errStack] = errInfo(err);
           _package.logger.error(errMsg, undefined, errStack);
@@ -110,10 +114,11 @@ export class HelmToMolfileConverter {
   private getPolymerMolfile(
     helm: string,
     polymerGraph: string,
-    rdKitModule: RDModule
+    rdKitModule: RDModule,
+    monomerLib: IMonomerLib
   ): string {
     const globalPositionHandler = new GlobalMonomerPositionHandler(polymerGraph);
-    const polymer = new Polymer(helm, rdKitModule);
+    const polymer = new Polymer(helm, rdKitModule, monomerLib);
     globalPositionHandler.monomerSymbols.forEach((monomerSymbol: string, monomerIdx: number) => {
       const shift = globalPositionHandler.getMonomerShifts(monomerIdx);
       polymer.addMonomer(monomerSymbol, monomerIdx, shift);