+++ packages/Bio/src/utils/monomer-lib/monomer-lib.ts
@@ -1,12 +1,13 @@
 /* Do not change these import lines to match external modules in webpack configuration */
 import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
+import * as DG from 'datagrok-api/dg';
 
 import wu from 'wu';
 import {Observable, Subject} from 'rxjs';
 
 import {
-  IMonomerLib, Monomer, MonomerType, PolymerType, RGroup
+  IMonomerLib, Monomer, MonomerLibSummaryType, MonomerType, PolymerType, RGroup
 } from '@datagrok-libraries/bio/src/types';
 import {
   HELM_REQUIRED_FIELD as REQ, HELM_RGROUP_FIELDS as RGP
@@ -161,12 +162,26 @@ export class MonomerLib implements IMonomerLib {
     this._onChanged.next();
   }
 
-  getSummary(): string {
-    const monTypeList: PolymerType[] = this.getPolymerTypes();
-    const resStr: string = monTypeList.length == 0 ? 'empty' : monTypeList.map((monType) => {
-      return `${monType} ${this.getMonomerSymbolsByType(monType).length}`;
-    }).join('\n');
-    return resStr;
+  getSummary(): MonomerLibSummaryType {
+    const res: MonomerLibSummaryType = {};
+    const ptList: PolymerType[] = this.getPolymerTypes();
+    for (const pt of ptList)
+      res[pt] = this.getMonomerSymbolsByType(pt).length;
+    return res;
+  }
+
+  getSummaryDf(): DG.DataFrame {
+    const ptList = this.getPolymerTypes();
+
+    const countList: number[] = new Array<number>(ptList.length);
+    for (const [pt, i] of wu.enumerate(ptList))
+      countList[i] = this.getMonomerSymbolsByType(pt).length;
+
+    const resDf: DG.DataFrame = DG.DataFrame.fromColumns([
+      DG.Column.fromStrings('polymerType', ptList),
+      DG.Column.fromList(DG.COLUMN_TYPE.INT, 'count', countList),
+    ]);
+    return resDf;
   }
 
   getTooltip(polymerType: PolymerType, monomerSymbol: string): HTMLElement {