+++ packages/Helm/src/package-test.ts
@@ -8,7 +8,7 @@ import './tests/helm-tests.ts';
 import './tests/findMonomers-tests';
 import './tests/helm-service-tests';
 import './tests/renderers-tests';
-import './tests/get-molfiles';
+import './tests/get-molfiles-tests';
 import './tests/properties-widget-tests';
 import './tests/get-all-parts-tests';
 import './tests/get-monomer-tests';
@@ -23,6 +23,7 @@ export {tests};
 //input: object testContext {optional: true}
 //output: dataframe result
 export async function test(category: string, test: string, testContext: TestContext): Promise<DG.DataFrame> {
-  const data = await runTests({category, test, testContext});
+  // verbose: true - for tests returning dataframe
+  const data = await runTests({category, test, testContext, verbose: true});
   return DG.DataFrame.fromObjects(data)!;
 }