+++ packages/Helm/src/utils/get-hovered.ts
@@ -2,12 +2,10 @@ import * as grok from 'datagrok-api/grok';
 import * as DG from 'datagrok-api/dg';
 import * as ui from 'datagrok-api/ui';
 
-import * as org from 'org';
 import * as JSDraw2 from 'JSDraw2';
 
-import {IMonomerLib} from '@datagrok-libraries/bio/src/types/index';
-import {IEditorMol, IEditorMolAtom} from '@datagrok-libraries/bio/src/types/helm-web-editor';
 import {helmTypeToPolymerType} from '@datagrok-libraries/bio/src/monomer-works/monomer-works';
+import {PolymerType} from '@datagrok-libraries/bio/src/types';
 
 import {HelmMonomerPlacer, ISeqMonomer} from '../helm-monomer-placer';
 
@@ -81,8 +79,9 @@ export function getHoveredMonomerFallback(
     }
   }
   left = (argsX >= sumLengths[left]) ? left : left - 1; // correct left to between sumLengths
-  if (left >= 0)
-    hoveredSeqMonomer = getSeqMonomerFromHelm(allParts[0], allParts[left], helmPlacer.monomerLib);
+  if (left >= 0) {
+    hoveredSeqMonomer = getSeqMonomerFromHelm(allParts[0], allParts[left]);
+  }
   return hoveredSeqMonomer;
 }
 
@@ -92,10 +91,11 @@ function getSeqMonomerFromHelmAtom(atom: JSDraw2.IEditorMolAtom): ISeqMonomer {
 }
 
 function getSeqMonomerFromHelm(
-  helmPrefix: string, symbol: string, monomerLib: IMonomerLib
+  helmPrefix: string, symbol: string
 ): ISeqMonomer {
   let resSeqMonomer: ISeqMonomer | undefined = undefined;
-  for (const polymerType of monomerLib.getPolymerTypes()) {
+  const polymerTypeList: PolymerType[] = ['RNA', 'PEPTIDE', 'CHEM', 'BLOB', 'G'];
+  for (const polymerType of polymerTypeList) {
     if (helmPrefix.startsWith(polymerType))
       resSeqMonomer = {symbol: symbol, polymerType: polymerType};
   }